// src/product/modules/product-detail/enum/frame.type.ts
export enum FrameType {
  FULL_RIM = 'full_rim',
  HALF_RIM = 'half_rim',
  NO_RIM = 'no_rim',
  RIMLESS = 'rimless',
}

export enum FrameShapeType {
  //Round, Square, Rectangle, Browline, Butterfly/Cat Eye, Aviator, Narrow, Oval
  ROUND = 'round',
  SQUARE = 'square',
  RECTANGLE = 'rectangle',
  BROWLINE = 'browline',
  BUTTERFLY = 'butterfly',
  AVIATOR = 'aviator',
  NARROW = 'narrow',
  OVAL = 'oval',
}

export enum FrameStyleType {
  // Classic, Modern, Extravagant, Sporty
  CLASSIC = 'classic',
  MODERN = 'modern',
  EXTRAVAGANT = 'extravagant',
  SPORTY = 'sporty',
}

export enum FrameBridgeType {
  // Without nose pads, With nose pads, With keyhole bridge
  WITHOUT_NOSE_PADS = 'without_nose_pads',
  WITH_NOSE_PADS = 'with_nose_pads',
  WITH_KEYHOLE_BRIDGE = 'with_keyhole_bridge',
}

export enum FrameMaterialType {
  // Plastic, Metal, Titan, Wood, Carbon, Aluminium, Cellulose, Leather
  PLASTIC = 'plastic',
  METAL = 'metal',
  TITAN = 'titan',
  WOOD = 'wood',
  CARBON = 'carbon',
  ALUMINIUM = 'aluminium',
  CELLULOSE = 'cellulose',
  LEATHER = 'leather',
}